+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## Unreleased
+## 3.30.0
 * Add `oauthAccessRevocation` to `WebhookNotification`s
 * Add support for `profileId` in Transaction#create options for VenmoAccounts
 * Add support for dispute search by `customerId`, `disbursementDate`, and `effectiveDate`
+* Make `CustomerGateway::find` backward compatible
 
 ## 3.29.0
 * Add support for `association_filter_id` in `Customer#find`